refactor : 클린 아키텍처 구조 변경
